py-scraper/requirements.txt
curl_cffi>=0.6.2
urllib3>=2.0.7
fastapi==0.109.2
uvicorn==0.24.0
pydantic>=2.6.0
python-dotenv==1.0.1
fake-useragent>=1.4.0
tweety-ns==2.1 